ec2-s3-CURD-project/requirements.txt
Flask==3.0.0
boto3==1.34.0
botocore==1.34.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.12
requests==2.31.0

